+++ src/main/java/codesquad/springcafe/service/ArticleService.java
@@ -1,9 +1,12 @@
 package codesquad.springcafe.service;
 
 import codesquad.springcafe.domain.Article;
+import codesquad.springcafe.domain.Reply;
 import codesquad.springcafe.dto.ArticleDto;
+import codesquad.springcafe.error.exception.AccessDeniedException;
 import codesquad.springcafe.error.exception.ArticleNotFoundException;
 import codesquad.springcafe.repository.article.ArticleRepository;
+import codesquad.springcafe.repository.reply.ReplyRepository;
 import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,10 +17,12 @@ import org.springframework.stereotype.Service;
 public class ArticleService {
     private static final Logger logger = LoggerFactory.getLogger(ArticleService.class);
     private final ArticleRepository articleRepository;
+    private final ReplyRepository replyRepository;
 
     @Autowired
-    public ArticleService(ArticleRepository articleRepository) {
+    public ArticleService(ArticleRepository articleRepository, ReplyRepository replyRepository) {
         this.articleRepository = articleRepository;
+        this.replyRepository = replyRepository;
     }
 
     public void createArticle(Article article) {
@@ -48,6 +53,16 @@ public class ArticleService {
     }
 
     public void deleteArticle(long articleId) {
+        Article article = articleRepository.findByArticleId(articleId)
+                .orElseThrow(() -> new ArticleNotFoundException(articleId + " ID 게시글이 존재하지 않습니다."));
+
+        List<Reply> replies = replyRepository.findRepliesByArticleId(articleId);
+
+        if ((!replies.isEmpty()) && replies.stream()
+                .anyMatch(reply -> !reply.getWriter().equals(article.getWriter()))) {
+            throw new AccessDeniedException(articleId + " ID 게시글 삭제 불가");
+        }
+
         articleRepository.deleteArticle(articleId);
         logger.debug("ID {} 게시글 삭제", articleId);
     }
